Chore: 일부 구현 요소 수정 - RGB init 범위 예외 요소 구현 - Factory 기능 분할 - 원시 타입 직접 전달 최소화
